Move the `tags` pages under the blog directory.
The tests ended up being really useful for this.
