+++ src/pages/blog/index.tsx
@@ -2,10 +2,10 @@ import * as React from "react";
 import { PageProps, graphql } from "gatsby";
 import Link from "gatsby-link";
 
-import { BaseLayout } from "../layouts";
-import SEO from "../components/seo";
+import { BaseLayout } from "../../layouts";
+import SEO from "../../components/seo";
 
-import { BlogIndexPageQuery } from "../../graphql-types";
+import { BlogIndexPageQuery } from "../../../graphql-types";
 
 const BlogIndex = ({ data }: PageProps<BlogIndexPageQuery>): JSX.Element => {
   const posts = data.allBlogPosts.edges;
